+++ src/ripple/protocol/Protocol.h
@@ -95,7 +95,10 @@ using LedgerIndex = std::uint32_t;
 */
 using TxID = uint256;
 
-std::size_t constexpr maxDeletableAMMTrustLines = 512;
+/** The maximum number of trustlines to delete as part of AMM account
+ * deletion cleanup.
+ */
+std::uint16_t constexpr maxDeletableAMMTrustLines = 512;
 
 }  // namespace ripple
